# Updated file edit method

+++ cURL.py
@@ -1,7 +1,7 @@
 import argparse
 import os
-import os
-from termcolor import colored
+import os.path as os_path
+# from termcolor import colored
 
 # To make into .exe
 # Usage: --<HTTP request type> <request_route>
@@ -39,6 +39,16 @@ def process_yes_no(text):
     return False
 
 
+def open_txt_file(txt_file_name):
+    os.startfile(txt_file_name)
+    os.system('pause')
+
+
+def read_txt_file(txt_file_name):
+    f = open(txt_file_name, 'r')
+    return f.read().replace(' ', '')
+
+
 # vars() to make it iterable
 args = vars(parser.parse_args())
 
@@ -55,7 +65,6 @@ for req_type in req_types:
         http_request_type = req_type
         req_url = f'{url}{args[req_type]}'
 
-print(args)
 
 if(http_request_type == None):
     # select_req_type = colored('Please select a request type\n', 'red')
@@ -64,16 +73,45 @@ if(http_request_type == None):
     exit()
 
 if(http_request_type == 'post' and args['data'] == None):
-    yes_no = input('Send with request body data (Y/N)? \n')
-    with_req_body = process_yes_no(yes_no)
+    with_req_body_option = input('Send with request body data (Y/N)? \n')
+    with_req_body = process_yes_no(with_req_body_option)
 
     if(with_req_body):
-        f = open("req_body.txt", "w+")
-        os.startfile(f.name)
+
+        if(not os_path.exists('req_body.txt')):
+            f = open('req_body.txt', 'w+')
+
+            open_txt_file('req_body.txt')
+
+        else:
+            edit_now_option = input(
+                'Would you like to edit the contents now (Y/N)? \n')
+            edit_now = process_yes_no(edit_now_option)
+
+            if(edit_now):
+                open_txt_file('req_body.txt')
+
+        req_body_data = read_txt_file('req_body.txt')
+
+        if(req_body_data != ''):
+            print('\n')
+            print(req_body_data)
+
 
 else:
     try:
-        os.startfile(args['data'])
+        edit_now_option = input(
+            'Would you like to edit the contents now (Y/N)? \n')
+        edit_now = process_yes_no(edit_now_option)
+
+        if(edit_now):
+            open_txt_file(args['data'])
+
+            req_body_data = read_txt_file(args['data'])
+
+            if(req_body_data != ''):
+                print('\n')
+                print(req_body_data)
 
     except:
         print('Please input a valid file location')
